[~] renamed package "ui" to "presentation" and class "Hero" to "HeroUi"
