+++ app/src/main/java/com/example/marvelheroes/data/repository/HeroListRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.example.marvelheroes.data.mappers.toHeroEntity
 import com.example.marvelheroes.data.mappers.toHeroUi
 import com.example.marvelheroes.data.network.MarvelApi
 import com.example.marvelheroes.domain.repository.HeroListRepository
-import com.example.marvelheroes.ui.Hero
+import com.example.marvelheroes.presentation.HeroUi
 import com.example.marvelheroes.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
@@ -16,7 +16,7 @@ class HeroListRepositoryImpl @Inject constructor(
     private val marvelApi: MarvelApi,
     private val heroDatabase: HeroDatabase
 ): HeroListRepository {
-    override suspend fun getHeroList(forceFetchFromRemote: Boolean): Flow<Resource<List<Hero>>> {
+    override suspend fun getHeroList(forceFetchFromRemote: Boolean): Flow<Resource<List<HeroUi>>> {
         return flow {
 
             val localHeroList = heroDatabase.heroDao().getAllHeroes()
@@ -46,7 +46,7 @@ class HeroListRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getHero(id: String): Flow<Resource<Hero>> {
+    override suspend fun getHero(id: String): Flow<Resource<HeroUi>> {
         return flow {
             val localHeroList = heroDatabase.heroDao().getAllHeroes()
