+++ app/src/main/java/com/example/marvelheroes/presentation/mainscreen/HeroListState.kt
@@ -1,0 +1,9 @@
+package com.example.marvelheroes.presentation.mainscreen
+
+import com.example.marvelheroes.presentation.HeroUi
+
+data class HeroListState(
+    val isLoading: Boolean = false,
+    val isError: Boolean = false,
+    val heroList: List<HeroUi> = emptyList()
+)
