+++ app/src/main/java/com/example/marvelheroes/presentation/secondscreen/HeroState.kt
@@ -1,0 +1,8 @@
+package com.example.marvelheroes.presentation.secondscreen
+
+import com.example.marvelheroes.presentation.HeroUi
+
+data class HeroState(
+    val isError: Boolean = false,
+    val hero: HeroUi? = null
+)
