+++ app/src/main/java/com/example/marvelheroes/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.marvelheroes.ui.theme
+package com.example.marvelheroes.presentation.theme
 
 import androidx.compose.ui.graphics.Color
