--- app/src/main/java/com/example/marvelheroes/ui/secondscreen/HeroState.kt
@@ -1,8 +1,0 @@
-package com.example.marvelheroes.ui.secondscreen
-
-import com.example.marvelheroes.ui.Hero
-
-data class HeroState(
-    val isError: Boolean = false,
-    val hero: Hero? = null
-)
